move primitives from components
