+++ src/pages/about.tsx
@@ -1,7 +1,7 @@
 import type { ReactElement } from 'react'
 import { Layout } from '../components/layout'
 import { dict } from '../constant'
-import { H2, P } from '../components/typography'
+import { H2, P } from '../primitives/typography'
 
 export default function About(): ReactElement {
   return (
