+++ src/pages/config.tsx
@@ -1,8 +1,8 @@
 import type { FC } from 'react'
 import { useVolStatus } from '../api/use-vol-status'
-import { Button } from '../components/button'
+import { Button } from '../primitives/button'
 import { Layout } from '../components/layout'
-import { H3, Muted, P, Small } from '../components/typography'
+import { H3, Muted, P, Small } from '../primitives/typography'
 import { defaultConfig, useConfig } from '../config/use-config'
 import { dict } from '../constant'
 import { Input } from '../primitives/input'
